Added delete product modal to confirm to admin user that product will be deleted

+++ products/views.py
@@ -143,6 +143,16 @@ def delete_product(request, product_id):
         return redirect(reverse('home'))
 
     product = get_object_or_404(StoreProduct, pk=product_id)
-    product.delete()
-    messages.success(request, 'Product deleted!')
-    return redirect(reverse('products'))
+    if request.method == 'POST':
+        product.delete()
+        messages.add_message(
+            request, messages.SUCCESS,
+            'Product deleted!'
+        )
+        return redirect('products')
+
+    template = 'products/delete_product.html'
+    context = {
+        'product': product,
+        }
+    return render(request, template, context)
